chore: renamed new files to follow the naming convention
