+++ core/src/commands/base.ts
@@ -19,7 +19,7 @@ import type { Log } from "../logger/log-entry.js"
 import type { LoggerType, LoggerBase, LoggerConfigBase, LogLevel } from "../logger/logger.js"
 import { eventLogLevel } from "../logger/logger.js"
 import { printFooter } from "../logger/util.js"
-import { getCloudDistributionName, getCloudLogSectionName } from "../util/garden-cloud.js"
+import { getCloudDistributionName, getCloudLogSectionName } from "../util/cloud.js"
 import { getDurationMsec, getPackageVersion, userPrompt } from "../util/util.js"
 import { renderOptions, renderCommands, renderArguments, cliStyles, optionsWithAliasValues } from "../cli/helpers.js"
 import type { GlobalOptions, ParameterValues, ParameterObject } from "../cli/params.js"
