+++ core/src/commands/cloud/secrets/secrets-list.ts
@@ -18,7 +18,7 @@ import type { SecretResult } from "../helpers.js"
 import { applyFilter, makeSecretFromResponse, noApiMsg } from "../helpers.js"
 import { sortBy } from "lodash-es"
 import { StringsParameter } from "../../../cli/params.js"
-import { getCloudDistributionName } from "../../../util/garden-cloud.js"
+import { getCloudDistributionName } from "../../../util/cloud.js"
 import type { CloudApi, CloudProject } from "../../../cloud/api.js"
 import type { Log } from "../../../logger/log-entry.js"
 import { styles } from "../../../logger/styles.js"
