+++ core/src/commands/login.ts
@@ -19,7 +19,7 @@ import type { EventBus } from "../events/events.js"
 import type { ProjectConfig } from "../config/project.js"
 import { findProjectConfig } from "../config/base.js"
 import { BooleanParameter } from "../cli/params.js"
-import { getCloudDistributionName } from "../util/garden-cloud.js"
+import { getCloudDistributionName } from "../util/cloud.js"
 import { deline } from "../util/string.js"
 
 const loginTimeoutSec = 60
