+++ core/src/commands/logout.ts
@@ -10,7 +10,7 @@ import type { CommandParams, CommandResult } from "./base.js"
 import { Command } from "./base.js"
 import { printHeader } from "../logger/util.js"
 import { CloudApi, getGardenCloudDomain } from "../cloud/api.js"
-import { getCloudDistributionName } from "../util/garden-cloud.js"
+import { getCloudDistributionName } from "../util/cloud.js"
 import { dedent, deline } from "../util/string.js"
 import { ConfigurationError } from "../exceptions.js"
 import type { ProjectConfig } from "../config/project.js"
