+++ core/src/commands/serve.ts
@@ -14,7 +14,7 @@ import { printEmoji, printHeader } from "../logger/util.js"
 import { dedent } from "../util/string.js"
 import type { CommandLine } from "../cli/command-line.js"
 import { GardenInstanceManager } from "../server/instance-manager.js"
-import { getCloudDistributionName } from "../util/garden-cloud.js"
+import { getCloudDistributionName } from "../util/cloud.js"
 import { sleep } from "../util/util.js"
 import type { Log } from "../logger/log-entry.js"
 import { findProjectConfig } from "../config/base.js"
