+++ core/src/garden.ts
@@ -28,7 +28,7 @@ import {
   defaultNamespace,
   defaultEnvironment,
 } from "./config/project.js"
-import { getCloudDistributionName, getCloudLogSectionName } from "./util/garden-cloud.js"
+import { getCloudDistributionName, getCloudLogSectionName } from "./util/cloud.js"
 import { findByName, pickKeys, getPackageVersion, getNames, findByNames, duplicatesByKey } from "./util/util.js"
 import type { GardenError } from "./exceptions.js"
 import {
